ENH: Adding Theme Selection Value to Overlay
- These are import fixes/updates and splitting out the Cookies endpoint
(POST) to be a blueprint.
- This also sets-up and implements the feature such that now whenever
a theme cookie is set the dropdown menu will choose the current cookie
value as the currently selected theme. This makes the UX a little nicer
so that when you go to pick a new theme you're seeing your current
selection first (and with the checkmark) and you know the opposite is
the new one.
- Way less confusing, much much nicer experience.